refactored requirements-run, working on pruning
requirements-run now accepts an optional bound
and handles all the binding for distance functions
and the like if the bound > 1.

Realized that would could efficiently prune
runs for insensitive items if we do large spans
and work outside in.  If values don't change between
the bounds, we know they won't change for finer
grained values so those can be pruned.

This should help runtime substantially.