Implemented sparse sensitivity analysis, refined
Ported original naive sensitivity analaysis
for CMDD, which allowed us to get a decent
response but at significant runtime cost.

Realizing that we have a monotonic function,
we can be smart about exploring the domain
and leverage piecewise linear relationships
where we infer them (based on sampling).

That is, if we have an insensitivity in
the response between 2 sampled points,
we can assume any intermediate points
have the same value (horizonal line).

The search process then transforms from
naively evaluating each point to something
resembling a bisection search, where we
construct enveloping segments over a domain
and evaluate the extreme points.

Insensitivity indicates we can prune any
intermediate points, otherwise we bisect the
region into smaller segments and continue.

This alone nets a useful sparse sampling
benefit for very large finite domains
(notional data sees ~80% savings in
computation).  We can still improve
upon the method if the caller supplies
discrete samples.  The default extrema
search will end up over sampling
since it's dutifully splitting segments
equally.  If instead we let the requested
sampling data inform the creation of segments,
the technique scales down to smaller discrete
domains, while affording better than or equal
to performance on large fine grained domains.

When segmenting, we now partition the unvisited
user-specified sampling points into the segments,
and then shrink the segment such that the
end points are defined as the minimum/maximum
of the unvisited set.  This ensures we never
sample from intermediate points that are outside
the requested discrete domain, so that in the
worst case, we only ever sample the exact points
requested.  In the best case, we sample fewer than
requested and are able to prune the input substantially.