Fixed error in expand-misses, working edge case
expand-misses was uninentionally creating duplicate
records, throwing off the bound information, and
sampling from the wrong record to duplicate.  Instead
of drawing from the end, it drew earlier due to improper
application of laziness.  Fixing that brings
the faster variant in line with the naive requiremnts
contour variant.

The pruned variant has an inexplicable edge case at
cMDD of 2, where the pruned variant believes that 74,
a net of 39 supply each, is sufficient to meet the
2 day cutoff.  This appears valid when run via
history->contiguous-misses with a marathon stream.

However we're not getting similar behavior in the existing
pipeline, and instead have large blocks of requirement
that are the same (e.g. from 0 to 14 or so I think).

Could be a bug in the scoring function or something
else.