Case dive to determine small cmmd artifact
Derived a fixed set of cases for 01605K100
which displays a discrepancy between
the naive, faster, and pruned approaches.

Objective values change for CMDD of 2,
only when looking over the input CMDD
set of [0 2 ...].  Assuming 2 is not
a lower bound, we see results for
naive/faster that are equivalent along
the sample path during searching:

both go through [growth, misses(growth)]
points of:

[72 11] [76 14] [77 14] [78 0] [80 0]
[96 0] (in ascending order for clarity)

It turns out, that at CMDD of 2, there's
a point at [74 0] that is missed since
we assume that requirements analysis
is monotonic w.r.t. "some" requirement,
or in the case of CMDD "some" contiguous
missed demand > than the threshold.

When analyzing CMDD misses for a supply
of 39 (this unit type has only reserve
compos, both are identical in size),
at CMDD of 0, we get a CMDD missed of
at least 2. At CMDD of 2, we pass.

So the bisection search that the naive
and faster approaches are performing
is overlooking this point.

The pruning search incorporates previous
information from lower/higher bounds to
tighten the bounds and engages in a
different search trajectory.  It ends
up sampling 74, which gives a supply of
39, which yields the point by happenstance.

This looks like a unique phenomenon
exposing some sensitivity toward timing
and supply distribution that manifests
at the day(s) scale for "some" SRCs.
In this case, it appears to be RC-only,
so existing methodology for the AC may be
unaffected.

Results for this unit type are insensitive
to prefill.

On the balance: all 3 methods are leveraging
information from the CMDD requirement to
determine bounds to execute a bisection search
over an assumably monotonic function.  In
some apparently rare circumstances, the function
is not purely monotonic, so we have "hidden"
minima that may elude a bisection search depending
on the search trajectory.

The pruned variant may deviate (ideally it should
for efficiency gains) from the trajectories
that the naive/faster methods take by leveraging
previous samples to establish tighter bounds
based on the assumption that missed demands
as a function CMDD is similarly monotonic (this
looks solid).  In some cases this altered
trajectory may end up finding a minimum that
the bisection search - intialized with different,
wider bounds - missed.