Fix bugs for separating forward stationed units
If we want to tag unit entities with tags from Tags in supply
records, we need to add those to another field in supply records
called :unit-tags.

Forward needed to be added to the NonBOG categories so that we
don't trip the ineligible deployer exception.

Needed to add NonBOG computed supply to the Forward category.