Fixed typo and bug
Needed to let an empty table be returned from the transform
in case one of the threads throws an error and an empty table
is returned.

Also added a failing test to check if the new changes are
consistent with the outputs before these changes since the
outputs should match.  I think I forgot to use one thread
when creating the historical output data though..  Hopefully
I'm about to fix that.