Minor mod to supply tracking for compatibility.
Updated marathon.ces.supply/add-deployable-supply
to assoc the unit name instead of the unit entity
into what was an {category {src {nm entity}}}
mapping.

We now have {category {src {nm nm}}} which seems
redunant but it's fine.  The only downstream
traces of note were in find-feasible-supply,
which provides its own nm->unit function
ends up being get-entity closed over the
current context.  So we appear to be
okay, all tests passing, and now
the mutable store is consistent with the
persistent one on at least the first day.

Prior to this, we had a percieve inconsistency
since the entity being stored in the
SupplyStore {:deployable-buckets {:default {
"SRC3" {"38_SRC3_AC" ent}} was
a persistent spork.entitysystem.store.entity.

This differed from the (mutable) entity
under the mutable store, which continued
to mutate underneath while the entity
reference was store in 2 (or more) places.

Bottom line is we probably want to have some
kind of gross warning to ensure we are NEVER
associng a mentity as a value anywhere (or
an entity) unless we are certain of its
niche scope and lifecycle or we are
dumping it into a persistent entity.