Spec entity maps for random data replacement
This was more general as opposed to generating random
values for each of the fields based on their predefined
specs.  Using a spec entity map allows us to compose multiple
spec entity maps when two or more fields have data dependencies
within one record like the generator for Category and
DemandGroup where we ensure that the Category is the same for
all records of one DemandGroup.