Changed logging API in marathon.analysis.random
marathon.analysis.util/log now delegates to an
atom dereference for a logging function ::
any -> nil.  The default implementation is
println.  This retains legacy behavior of
echoing output to *out*.

Now we have a couple of new macros that
help us to log.

with-log-fn is a generic way to create
a scoped context where the value for
util/log is set to a different function

could probably do this as easily / better
with bindings, but I think atoms deref
much faster so meh.

building on this, util/log-to
implements the old log-to bevahior,
but for a scoped context.  So you
provide it path (e.g. a filename),
and the logger will redirect
output to that path, closing
the writer when the body is complete.

Note: thes are expected to wrap
eager/forced operations like a
pipeline doing work.  Lazy stuff
will have the same drawbacks as
with-open e.g. the lazy seq will
be created and unforced.